Reorganize RS485 Mode - FlowCtrl
